requirements.txt
anthropic>=0.40.0
httpx>=0.27.0
faster-whisper>=1.0.0
flask>=3.0.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.120.0
